+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2025-08-05 03:29:31
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: 874d4a02959d1f7d42c2d7640c0e4f8470b28451
+
+## [ForwardWidgets-2kuai] - 2025-08-05 03:29:31
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: c85dc8172e61048b4a927c9df03841a057016097
+
+## [ForwardWidgets-pack1r] - 2025-08-05 03:29:31
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: 5a8b6b71ddd0d6e4321c8c1851dd9be229179666
+
+
 ## [ForwardWidgets-huangxd] - 2025-08-04 03:47:04
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main